Feature: Support for add non form controls.
